Change lock

+++ bin/scripts/dead_lock.sh
@@ -1,9 +1,16 @@
 #!/bin/sh
 
-~/.my_bin/epi3lock -i ~/wallpaper/lock2.png &
+source ~/.export.sh
+
+nohup ~/.my_bin/epi3lock -i $LOCKSCREEN &
+
 pid=$!
-$CONFIG_GIT_DIR/bin/scripts/dead_clock.sh &
-pidbis=$!
+
+if [ "$DEADLOCK" = "true" ]; then
+    $CONFIG_GIT_DIR/bin/scripts/dead_clock.sh &
+    pidbis=$!
+    wait $pid
+    read
+    kill $pidbis
+fi
 wait $pid
-read
-kill $pidbis
